Change from usize to u64 (#29)
* Change from usize to u64

On [random-access-storage](https://github.com/datrs/random-access-storage)
issue
https://github.com/datrs/random-access-storage/issues/6 has changed all
types from `usize` into `u64` to be able to handle more than 4gbs on
32bits systems.

> usize is 32 bits on a 32 bit system, so the storage would be limited
to 4GB on such a system.

When changing `random-access-storage` on `hypercore` (tracking:
https://github.com/datrs/hypercore/pull/100) one of the things I've
noticed is that it wold be easier to also make `merkle-tree-stream` use
`u64` to integrate with `random-access-storage`. Very likely, this also
means that we would need to bump to use more than 32Gb storages on
`merkle-tree-stream`.

I've simply changed all declarations of usize into u64. All tests are
passing.

Needs:
---

- [ ] https://github.com/datrs/flat-tree/pull/44

* Fix travis

+++ src/lib.rs
@@ -89,13 +89,13 @@ pub trait HashMethods {
 /// works with.
 pub trait Node {
   /// Get the length of the node.
-  fn len(&self) -> usize;
+  fn len(&self) -> u64;
   /// Check if the length is zero.
   fn is_empty(&self) -> bool;
   /// Get the position of the parent of the node.
-  fn parent(&self) -> usize;
+  fn parent(&self) -> u64;
   /// Get the position at which the node was found.
-  fn index(&self) -> usize;
+  fn index(&self) -> u64;
   /// Get the hash contained in the node.
   fn hash(&self) -> &[u8];
 }
@@ -178,7 +178,7 @@ pub trait Node {
 pub struct MerkleTreeStream<T: HashMethods> {
   handler: T,
   roots: Vec<Arc<T::Node>>,
-  blocks: usize,
+  blocks: u64,
 }
 
 impl<H: HashMethods> MerkleTreeStream<H> {
@@ -194,13 +194,13 @@ impl<H: HashMethods> MerkleTreeStream<H> {
   /// Pass a string buffer through the flat-tree hash functions, and write the
   /// result back out to "nodes".
   pub fn next<'a>(&mut self, data: &[u8], nodes: &'a mut Vec<Arc<H::Node>>) {
-    let index: usize = 2 * self.blocks;
+    let index: u64 = 2 * self.blocks;
     self.blocks += 1;
 
     let leaf = PartialNode {
       index,
-      parent: flat::parent(index) as usize,
-      length: data.len(),
+      parent: flat::parent(index) as u64,
+      length: data.len() as u64,
       data: NodeKind::Leaf(data.to_vec()),
     };
 
@@ -223,7 +223,7 @@ impl<H: HashMethods> MerkleTreeStream<H> {
         let hash = self.handler.parent(left, right);
         let partial = PartialNode {
           index: left.parent(),
-          parent: flat::parent(left.parent()) as usize,
+          parent: flat::parent(left.parent()) as u64,
           length: left.len() + right.len(),
           data: NodeKind::Parent,
         };
